+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ###v0.2.0-alpha on t2s2-
 - adding AWS SDK
+- new attributes for User:
+  - location_name (string) (max 128) until we have a proper location handling
+  - device_token (string) (max 128) ios device token for push notifications
 
 ###v0.1.3-alpha on t2s2-
 Social Connect: